requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
discord.py==2.3.2
redis==4.5.4
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic==1.10.7
aiohttp==3.8.4
cryptography==39.0.2 